--- src/main/java/Main.java
@@ -1,8 +1,0 @@
-public class Main {
-
-    public static void main(String[] args) {
-        MessagePrinter messagePrinter = new MessagePrinter();
-        messagePrinter.printMessage();
-    }
-
-}
