--- src/main/java/MessegePrinter.java
@@ -1,8 +1,0 @@
-class MessagePrinter {
-    private final SimpleMessageProducer messageProducer = new SimpleMessageProducer();
-
-    public void printMessage() {
-        String message = messageProducer.getMessage();
-        System.out.println(message);
-    }
-}
