--- src/main/java/SimpleMessegeProducer.java
@@ -1,5 +1,0 @@
-class SimpleMessageProducer {
-        String getMessage() {
-            return "Example message " + System.currentTimeMillis();
-        }
-    }
