+++ src/main/java/com/example/Main.java
@@ -1,0 +1,15 @@
+package com.example;
+
+import com.example.test.MessagePrinter;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.ConfigurableApplicationContext;
+
+@SpringBootApplication
+public class Main {
+    public static void main(String[] args) {
+        ConfigurableApplicationContext context = SpringApplication.run(Main.class);
+        MessagePrinter messagePrinter = context.getBean(MessagePrinter.class);
+        messagePrinter.printMessage();
+    }
+}
