+++ src/main/java/com/example/test/MessageProducer.java
@@ -1,0 +1,5 @@
+package com.example.test;
+
+public interface MessageProducer {
+    String getMessage();
+}
