--- src/main/java/com/example/test/TestApplication.java
@@ -1,13 +1,0 @@
-package com.example.test;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TestApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(TestApplication.class, args);
-    }
-
-}
